Add comprehensive unit tests for SearchEngine functionality
- Implemented a new test suite for the SearchEngine class, covering various aspects including initialization, document management, search functionality, caching, background processing, and performance metrics.
- Added tests for basic search operations, edge cases, incremental search features, and signal emissions.
- Included tests for handling special characters and Unicode searches.
- Ensured proper cleanup and resource management in test cases.

+++ tests/search/memory_manager_test.cpp
@@ -386,7 +386,7 @@ MemoryManager::MemoryStats MemoryManagerTest::createMockStats()
     stats.textCacheMemory = 1024 * 1024 * 15; // 15MB
     stats.highlightCacheMemory = 1024 * 1024 * 10; // 10MB
     stats.systemMemoryUsage = 1024 * 1024 * 1024; // 1GB
-    stats.systemMemoryTotal = 1024 * 1024 * 1024 * 4; // 4GB
+    stats.systemMemoryTotal = int64_t(1024) * 1024 * 1024 * 4; // 4GB
     stats.memoryPressure = 0.25;
     stats.pressureLevel = MemoryManager::Normal;
     stats.optimizationCount = 5;
